office-updates-04052026

+++ CbsAp/CbsAp.Application/Features/KeywordManagement/Command/DeleteKeyword/DeleteKeywordCommandHandler.cs
@@ -3,6 +3,7 @@ using CbsAp.Application.Abstractions.Persistence;
 using CbsAp.Application.Abstractions.Services.Shared;
 using CbsAp.Application.Configurations.constants;
 using CbsAp.Application.Shared.ResultPatten;
+using CbsAp.Domain.Entities.Invoicing;
 using CbsAp.Domain.Entities.Keywords;
 using CbsAp.Domain.Enums;
 using Microsoft.Extensions.Logging;
@@ -32,10 +33,16 @@ namespace CbsAp.Application.Features.KeywordManagement.Command
             var keywordToDelete = await keywordRepo
                 .SingleOrDefaultAsync(e => e.KeywordID == request.KeywordID);
 
+            var invoiceRepo = _unitofWork.GetRepository<Invoice>();
+            var isKeywordActive = await invoiceRepo.AnyAsync(a => a.KeywordID == request.KeywordID);
+
+            if (isKeywordActive)
+                return ResponseResult<bool>.BadRequest("This keyword is currently in use and cannot be deleted.");
+
             if (keywordToDelete != null)
             {
                 await keywordRepo.DeleteAsync(keywordToDelete);
-                await _unitofWork.SaveChanges(cancellationToken);
+                await _unitofWork.SaveChanges(string.Empty, string.Empty, cancellationToken);
             }
             
             return ResponseResult<bool>.OK(MessageConstants.Message("Keyword", MessageOperationType.Delete));
